passes: Refactor DataManagers
