+++ devito/passes/iet/languages/basic.py
@@ -12,11 +12,12 @@ from devito.ir import (DummyEq, Conditional, Dereference, Expression, Expression
 from devito.mpi.routines import IrecvCall, IsendCall
 from devito.symbolics import CondEq, INT, ccode
 from devito.passes.iet.engine import iet_pass
+from devito.passes.iet.misc import is_on_device
 from devito.tools import as_tuple, is_integer, prod
 from devito.types import PointerArray, Symbol, NThreadsMixin
 
 __all__ = ['Constructs', 'LanguageTransformer', 'PragmaSimdTransformer',
-           'PragmaShmTransformer', 'PragmaDeviceAwareTransformer', 'is_on_device']
+           'PragmaShmTransformer', 'PragmaDeviceAwareTransformer']
 
 
 class Constructs(dict):
@@ -550,7 +551,7 @@ class PragmaDeviceAwareTransformer(PragmaShmTransformer):
         for parallelism. In particular:
 
             * All parallel Iterations not *writing* to a host Function, that
-              is a Function `f` such that ``is_on_device(f) == False`, are offloaded
+              is a Function `f` such that `is_on_device(f) == False`, are offloaded
               to the device.
             * The remaining ones, that is those writing to a host Function,
               are parallelized on the host.
@@ -613,36 +614,3 @@ class PragmaDeviceAwareTransformer(PragmaShmTransformer):
         iet = Transformer(mapper).visit(iet)
 
         return iet, {}
-
-
-# Utils
-
-def is_on_device(maybe_symbol, gpu_fit, only_writes=False):  #TODO: MAKE IT CLASSMETHOD OF DEVICEPARALLELIZER ??
-    """
-    True if all given Functions are allocated in the device memory, False otherwise.
-
-    Parameters
-    ----------
-    maybe_symbol : Indexed or Function or Node
-        The inspected object. May be a single Indexed or Function, or even an
-        entire piece of IET.
-    gpu_fit : list of Function
-        The Function's which are known to definitely fit in the device memory. This
-        information is given directly by the user through the compiler option
-        `gpu-fit` and is propagated down here through the various stages of lowering.
-    only_writes : bool, optional
-        Only makes sense if `maybe_symbol` is an IET. If True, ignore all Function's
-        that do not appear on the LHS of at least one Expression. Defaults to False.
-    """
-    try:
-        functions = (maybe_symbol.function,)
-    except AttributeError:
-        assert maybe_symbol.is_Node
-        iet = maybe_symbol
-        functions = set(FindSymbols().visit(iet))
-        if only_writes:
-            expressions = FindNodes(Expression).visit(iet)
-            functions &= {i.write for i in expressions}
-
-    return all(not (f.is_TimeFunction and f.save is not None and f not in gpu_fit)
-               for f in functions)
